requirements.txt
streamlit
torch
torchvision
transformers
pillow
opencv-python
datasets
numpy
google-generativeai
